# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation for the Bookstore API",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/anggaran": {
            "get": {
                "tags": [
                    "Anggaran"
                ],
                "summary": "Ambil semua data anggaran",
                "operationId": "getAnggaranList",
                "responses": {
                    "200": {
                        "description": "Berhasil ambil data anggaran",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Anggaran"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Anggaran"
                ],
                "summary": "Tambah data anggaran",
                "operationId": "createAnggaran",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Anggaran"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil tambah data anggaran",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Anggaran"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/anggaran/{id}": {
            "get": {
                "tags": [
                    "Anggaran"
                ],
                "summary": "Tampilkan detail anggaran",
                "operationId": "getAnggaranById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil ambil data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Anggaran"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Anggaran"
                ],
                "summary": "Update data anggaran",
                "operationId": "updateAnggaran",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Anggaran"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil update data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Anggaran"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Anggaran"
                ],
                "summary": "Hapus data anggaran",
                "operationId": "deleteAnggaran",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Berhasil hapus"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Anggaran": {
                "title": "Anggaran",
                "required": [
                    "user_id",
                    "kategori_id",
                    "jumlah_anggaran",
                    "periode"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "kategori_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "jumlah_anggaran": {
                        "type": "integer",
                        "example": 1000000
                    },
                    "periode": {
                        "type": "string",
                        "example": "2025-01"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-19T08:30:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-19T08:30:00Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}